refactor: extract RangeTreeViewBase to consolidate duplicate progressive loading logic
